requirements.txt
flask
requests
openai==0.28
python-dotenv
streamlit